create flyout submission detail styles
